rename glancesync_wrapper to glancesync_serversfacade
